fix(30106): export defaults override
