+++ hivemq-edge/src/frontend/src/modules/Bridges/hooks/useBridgeConfig.tsx
@@ -1,6 +1,7 @@
 import type { Bridge } from '@/api/__generated__'
 import type { Dispatch, FunctionComponent, PropsWithChildren, SetStateAction } from 'react'
 import { createContext, useContext, useState } from 'react'
+import { bridgeInitialState } from '@/modules/Bridges/utils/defaults.utils.ts'
 
 interface BridgeContextProps {
   bridge: Bridge
@@ -8,18 +9,6 @@ interface BridgeContextProps {
 }
 
 export const BridgeContext = createContext<BridgeContextProps | null>(null)
-
-// TODO The number and booleans should all be coming from the openAPI specs since they are all mandatory (see $Bridge.properties.cleanStart.isRequired)
-// eslint-disable-next-line react-refresh/only-export-components
-export const bridgeInitialState: Bridge = {
-  cleanStart: true,
-  host: '',
-  keepAlive: 60,
-  id: '',
-  port: 1883,
-  sessionExpiry: 3600,
-  persist: true,
-}
 
 export const BridgeProvider: FunctionComponent<PropsWithChildren> = ({ children }) => {
   const [bridge, setBridge] = useState<Bridge>(bridgeInitialState)
